+++ templates/play.html
@@ -4,7 +4,7 @@
 
 <link rel="stylesheet" type= "text/css" href="{{url_for('static', filename='stylesheets/wikipedia.css')}}">
 <link rel="stylesheet" type= "text/css" href="{{url_for('static', filename='stylesheets/play.css')}}">
-<script type="module" src="{{url_for('static', filename='js/play wrapper.js')}}"></script>
+<script type="module" src="{{url_for('static', filename='js/play.js')}}"></script>
 
 {% endblock %}
